Refactor invite token processing into a concern
