+++ app/controllers/outputs_controller.rb
@@ -1,5 +1,6 @@
 class OutputsController < ApplicationController
   include CreatingMultipleStudyResources
+  include InvitingUsers
 
   before_action :set_study_from_study_id, only: [:new, :create]
   before_action :check_user_can_contribute_to_study, only: [:new, :create]
@@ -20,19 +21,6 @@ class OutputsController < ApplicationController
     else
       raise ActionController::RoutingError.new("Not Found")
     end
-  end
-
-  # Override render and redirect_to so that we sign the invited user out
-  # before we render or redirect anywhere.
-  def render(*args)
-    sign_out_invited_user
-    super
-  end
-
-  # This is a "just in case" - no invited user should be redirect anywhere
-  def redirect_to(*args)
-    sign_out_invited_user
-    super
   end
 
   private
@@ -108,26 +96,11 @@ class OutputsController < ApplicationController
                                         :publication_year, :lead_author)
   end
 
-  def check_user_can_contribute_to_study
-    if params[:token].present?
-      @invite_token = params[:token]
-      @invited_user = User.find_by_invite_token(params[:token])
-      return forbidden if @invited_user.blank?
-      @invite = StudyInvite.where(study: @study, invited_user: @invited_user)
-      return forbidden if @invite.blank?
-      # We have to sign in the user because we rely on the current_user helper
-      # in lots of places to tie up actions and outputs with users, however,
-      # we sign them out again after the action has finished so that they
-      # don't ever get access to things they shouldn't
-      sign_in(@invited_user)
-    else
-      check_user_can_manage_study
-    end
+  def set_invited_user
+    @invited_user = User.find_by_invite_token(@invite_token)
   end
 
-  def sign_out_invited_user
-    unless @invited_user.blank?
-      sign_out(@invited_user)
-    end
+  def set_invite
+    @invite = StudyInvite.where(study: @study, invited_user: @invited_user)
   end
 end
